pyproject.toml
[tool.poetry]
name = "homework v2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dates = "^0.0.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.2.2"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs=true
warn_return_any=true

[tool.black]
line-length = 119
exclude = '''
(
	/(
		\.eggs # Исключить несколько общих каталогов
		| \.git # в корне проекта
		| \.hg
		| \.mypy_cache
		| \.tox
		| \.venv
		| dist
	)/
	| foo.py # Также отдельно исключить файл с именем foo.py # в корне проекта
	)
'''

[tool.isort]
line_length = 119
